BDOG-2213 Update Vulnerabilities service to use summaries collection
1. By switching to use a collection which is already structured as
required by the frontend - this significantly improves performance.
Benchmark of 20 local executions with no filters applied showed:
	Old flat collection: ~1400ms response time
	New nested collection: ~410ms response time
(Both figures refer to backend only)

2. Note, an aggregation pipeline is still required, in order to
filter out occurrences that don't contain the team/service. But
pattern matching is used to only use a pipeline if needed, and
to use the least amount of steps required.