BAU - Refactor service and tests
The WhatsRunningWhereService has been removed, as all the methods it provides could
more appropriately be added as private methods to the updateVulnerabilitiesService.

Some methods have also been removed from the Vulnerabilities service to the updateVulnerabilitiesService.
This is because the Vulnerabilities service should deal with requests from end users, whereas
the udpateVulnerabilitiesService concerns itself with updating and refining the data to be stored in mongo.

Prior to this change, the responsibilities were blurred.